Plot maelstRom's imprinting detection results.

maelstRom_imprintplot plots the results of maelstRom's imprinting detection analysis. More specifically, it plots a histogram of the observed reference allele fraction, als well as the PMF of the fitted imprinted model, which is a binomial mixture model using a heterozygous peak with no variant bias (p-parameter = 0.5) but that is split up according to a fitted degree of imprinting (see imprinting_est for this fitting procedure and pmf_impr for the imprinted pmf). Optionally, the non-imprinted fit (so just a binomial mixture model without variant bias) is plotted as well, if plot_NoImpr is TRUE.

maelstRom_imprintplot(
  ref_counts,
  var_counts,
  allelefreq,
  impr,
  SE,
  inbr = 0,
  PlotCov = 50,
  plot_NoImpr = FALSE,
  SplitPeaks = FALSE,
  ImprCols = NULL,
  NoImprCol = NULL,
  wd_res = NULL,
  chr = "",
  position = "",
  gene = ""
)

Arguments

ref_counts

Numeric vector. Reference counts of the locus.

var_counts

Numeric vector. Variant counts of the locus.

allelefreq

Number. Allele frequency.

impr

Number. Degree of imprinting.

SE

Number. Sequencing error rate.

inbr

Number. Degree of inbreeding (default = 0).

PlotCov

Number. Even though the observed number of reads can vary per sample, we have to assume a single number of observed reads to plot imprinting PMFs. One can opt to pick a set value for this (the default is 50) or e.g. use the locus' median or mean coverage here, so as to best correspond to the observed data.

SplitPeaks

Logical. If TRUE, the PMFs of both homozygotes and the heterozygotes are plotted seperately and using different colors. If FALSE, one total imprinted PMF is plotted.

NoImprCol

String. Color for the non-imprinted PMF (also no allelic bias; heterozygous binomial p-parameter = 0.5). There's no option to split the peaks of this PMF as it would make the final figure way too clustered, so only one color value is required. A default color is used in case non is given as input.

wd_res

String. Working directory where plots are saved; if non is given, the plot itself (as a ggplot2 object) is returned instead.

chr

Number. Chromosome of the locus, to be used in naming the outpud file if a wd_res is given.

position

Number. Position of the locus, to be used in naming the outpud file if a wd_res is given.

gene

String. Gene the locus is part of, to be used in naming the outpud file if a wd_res is given; optional.

Plot_NoImpr

Logical. If TRUE, also plot the non-imprintend PMF (a binomial mixture model without variant bias i.e. binomial p-parameter of heterozygotes = 0.5)

ImrCols

String (vector). Colors for the plotted PMFs of the imprinted fit. Three values are required if SplitPeaks is TRUE, otherwise just one value is required. Default colors are used in case none are given as input.