Various implementations of the tumor-purity-accounting (negative) log likelihood (sum of completely correlated beta-binomials)

UNUSED. These different implementations of, in essence, the same log-likelihood (to be used in maximum likelihood estimation) use different parameters transformations in order to find out which one is optimal for numerical optimization.

TumPur_LogLik_CPP(
  optpars,
  ref_counts,
  var_counts,
  tumpur,
  weights,
  SCPthreshold,
  n = 0L,
  NumIntMethod = "Gregory",
  prec = 1e-04,
  Wvec = 0L,
  Nvec = 0L
)

TumPur_LogLik_CPP2(
  optpars,
  ref_counts,
  var_counts,
  tumpur,
  weights,
  SCPthreshold,
  n = 0L,
  NumIntMethod = "Gregory",
  prec = 1e-04,
  Wvec = 0L,
  Nvec = 0L
)

TumPur_LogLik_CPP2X(
  optpars,
  ref_counts,
  var_counts,
  tumpur,
  weights,
  SCPthreshold,
  n = 0L,
  NumIntMethod = "Gregory",
  prec = 1e-04,
  Wvec = 0L,
  Nvec = 0L
)

TumPur_LogLik_CPP3(
  q,
  a,
  b,
  ref_counts,
  var_counts,
  tumpur,
  weights,
  SCPthreshold,
  n = 0L,
  NumIntMethod = "Gregory",
  prec = 1e-04,
  Wvec = 0L,
  Nvec = 0L
)

TumPur_LogLik_CPP3X(
  q,
  a,
  b,
  qlim,
  ref_counts,
  var_counts,
  tumpur,
  weights,
  SCPthreshold,
  n = 0L,
  NumIntMethod = "Gregory",
  prec = 1e-04,
  Wvec = 0L,
  Nvec = 0L
)

TumPur_LogLik_CPP4(
  q,
  a,
  b,
  ref_counts,
  var_counts,
  tumpur,
  weights,
  SCPthreshold,
  n = 0L,
  NumIntMethod = "Gregory",
  prec = 1e-04,
  Wvec = 0L,
  Nvec = 0L
)

TumPur_LogLik_CPP_DB1(
  optpars,
  ref_counts,
  var_counts,
  tumpur,
  weights,
  SCPthreshold,
  n = 0L,
  NumIntMethod = "Gregory",
  prec = 1e-04,
  Wvec = 0L,
  Nvec = 0L
)

TumPur_LogLik_CPP_DB2(
  optpars,
  ref_counts,
  var_counts,
  tumpur,
  weights,
  SCPthreshold,
  n = 0L,
  NumIntMethod = "Gregory",
  prec = 1e-04,
  Wvec = 0L,
  Nvec = 0L
)

TumPur_LogLik_CPP_DB3(
  optpars,
  ref_counts,
  var_counts,
  tumpur,
  weights,
  SCPthreshold,
  n = 0L,
  NumIntMethod = "Gregory",
  prec = 1e-04,
  Wvec = 0L,
  Nvec = 0L
)

TumPur_LogLik_CPP2_10(
  optpars,
  ref_counts,
  var_counts,
  tumpur,
  weights,
  SCPthreshold,
  n = 0L,
  NumIntMethod = "Gregory",
  prec = 1e-04,
  Wvec = 0L,
  Nvec = 0L
)

TumPur_LogLik_CPP2_100(
  optpars,
  ref_counts,
  var_counts,
  tumpur,
  weights,
  SCPthreshold,
  n = 0L,
  NumIntMethod = "Gregory",
  prec = 1e-04,
  Wvec = 0L,
  Nvec = 0L
)

TumPur_LogLik_CPP2_1000(
  optpars,
  ref_counts,
  var_counts,
  tumpur,
  weights,
  SCPthreshold,
  n = 0L,
  NumIntMethod = "Gregory",
  prec = 1e-04,
  Wvec = 0L,
  Nvec = 0L
)

TumPur_LogLik_CPP2Y(
  optpars,
  ref_counts,
  var_counts,
  tumpur,
  weights,
  SCPthreshold,
  n = 0L,
  NumIntMethod = "Gregory",
  prec = 1e-04,
  Wvec = 0L,
  Nvec = 0L
)

Arguments

optpars

Numeric vector. To be optimized parameters, usually consisting of the beta-binomial alpha (a) and beta (b) parameter, and a q-parameters denoting their difference (multiplicative factor) in cases compared to controls (same pi assumed, different theta; complete correlation within a sample assumed as well).

ref_counts

Numeric vector. Vector of reference allele counts.

var_counts

Numeric vector. Vector of variant allele counts.

tumpur

Numeric vector. Vector of tumor purities (or an estimate thereof) at expression-level.

weights

Numeric vector. Weights to be used for returning a weighted log-likelihood (sample-wise multiplication of log-likelihood with these weights).

SCPthreshold

Number. Every sample's corresponding tumor purity is used to model binomial sampling of tumor- and control reads making up the tumor observations. More specifically, every possible sampling's probability is tracked, and its corresponding log-likelihood (of the sum of completely correlated beta-binomials) calculated. This might be a bit overkill however; when SCPthreshold is specified, only the most likely binomial read sampling are taken into account, up until their total binomial probability sums to SCPthreshold. When SCPthreshold is set to 0, only the single most likely sampling is taken into account.

n

Number. Number of intervals to be used during calculations of sum-of-completely-correlated-beta-binomial likelihood via numerical integration, when the used integration method (NumIntMethod) is either the modified trapezoidal method using Gregory weights ("Gregory") or Tanh-sinh quadrature ("TanhSinhQuad")

NumIntMethod

String. Used numeric integration method (see above); either the modified trapezoidal method using Gregory weights ("Gregory"), Tanh-sinh quadrature ("TanhSinhQuad"), Newton Cotes ("NewtonCotes"), or Gaussian quadrature ("GaussianQuad").

prec

Number. A low number (e.g. 0.0001) acting as stop criterion when NumIntMethod == "TanhSinhQuad" (its iterative calculations stops if the current calculation differs less then prec from the previous one).

Wvec

Numeric vector. Vector of numeric integration weights used when NumIntMethod == "NewtonCotes", or NumIntMethod == "GaussianQuad".

Nvec

Numeric vector. Vector of node values used when NumIntMethod == "GaussianQuad".

Value

Negative log-likelihood.