Combine p-values of SNPs per gene.

combine_p_gene combines per-SNP p-values into per-gene p-values; the method for doing so can be chosen via input

combine_p_gene(p_values, weights = NULL, method = "geometric", r = NULL)

Arguments

p_values

Numeric vector. P-values to combine.

weights

Numeric vector. Weights given to each observation (corresponds to how much an observation occurs, but can be non-integer as well).

method

String. Specifies the method for combining p-values, the options being "arithmetic" (mean), "geometric" (mean), "harmonic" (mean), "minimum" (p-value), "maximum" (p-value) or "GMP" (Generalized Mean P-value). In case of the latter, a value for r must be specified, and the GMP is then calculated as (sum(weights * p_values^r)/sum(weights))^(1/r), see (Wilson 2020) .

r

Numeric. r-value to be used in the calculation of the GMP, should that option be chosen as the method.

Value

Combined p-value of a gene.

References

Wilson DJ (2020). “Generalized mean p-values for combining dependent tests: comparison of generalized central limit theorem and robust risk analysis.” Wellcome Open Research, 5.

Examples

combine_p_gene(c(0.015, 0.414, 8.47E-03, 7.43E-03, 0.574, 0.837))
#> [1] 0.07567171
combine_p_gene(c(0.063, 0.725, 0.657, 0.378, 0.291))
#> [1] 0.3189523