Exact gradient of the beta-binomial log-likelihood function for pi using a multiprecision library.

grad_pi_MP calculates the value of the gradient of the beta-binomial log-likelihood function to pi at given data points, while avoiding numerical mistakes (catastrophic cancellations) due to extreme parameter values. This function is called by grad_pi if necessary, and should not be called outside of this.

grad_pi_MP(ms, ns, piX, thetaX, NecPres)

Arguments

ms

Numeric vector. Vector of number of successes

ns

Numeric vector. Vector of number of trials

piX

Number. Probability of success; 0 >= piX >= 1

thetaX

Number. Dispersion parameter; 0 >= thetaX > +Inf

NecPres

Number. Necessary Precision, i.e. number of bits, for an accurate gradient calculatation, as determined by the function grad_pi

Value

A numeric vector of the same length as ms and ns, containing the gradient to pi in the give data points