Estimate the degree of imprinting.

imprinting_est estimates the degree of imprinting using iterative likelihood ratio tests for a specific SNP, going from imprinting = 0 to imprinting = 1 in steps of 0.01, performing LRTs along the way (against imprinting = 0), and finally retaining the best fit. See the function pmf_impr for what the degree of imprinting entails, mathematically

Usage

imprinting_est(ref_counts, var_counts, allelefreq, SE, inbr = 0)

Arguments

ref_counts

Numeric list. Reference counts.

var_counts

Numeric list. Variant counts.

allelefreq

Number. Allele frequency.

SE

Number. Sequencing error rate.

inbr

Number. Degree of inbreeding (default = 0).

Value

A list containing the following components:

est_i

The estimated degree of imprinting.

LRT

The test statistic of the likelihood ratio test against no imprinting.

p_value

The p-value of the likelihood ratio test against no imprinting.

GOF_likelihood

A goodness-of-fit value based on count-corrected likelihood, i.e. mean log PMF-value according to pmf_impr multiplied by each sample's coverage + 1 across samples of a locus.

Examples

imprinting_est(c(5, 8, 10, 3, 5, 6, 23), c(8, 8, 6, 4, 4, 10, 0), 0.5, 0.002)
#> $est_i
#> [1] 0
#> 
#> $LRT
#> [1] 0
#> 
#> $p_value
#> [1] 1
#> 
#> $GOF_likelihood
#> [1] 0.389424
#> 
imprinting_est(c(5, 0, 0, 3, 5, 1, 23), c(1, 8, 6, 2, 0, 10, 0), 0.5, 0.002, 0.12)
#> $est_i
#> [1] 0.81
#> 
#> $LRT
#> [1] 4.60567
#> 
#> $p_value
#> [1] 0.0159332
#> 
#> $GOF_likelihood
#> [1] 0.3810867
#>