Make a plot summarizing allelic divergence (-related phenomena) across a chromosome

maelstRom_ADChromplot returns various ggplot-objects visualizing maelstRom's differential allelic divergence results across a chromosome, together with differential expression logFC which can be determined using the same RNAseq data, as well as (optionally) methylation- and copy-number-alteration-data if available. All of this data should be supplied as separate dataframes, which enables plotting even when some of this data is missing for some of the loci or genes.

maelstRom_ADChromplot(
  AD_Data,
  DE_Data,
  Meth_Data = NULL,
  CNAgain_Data = NULL,
  CNAloss_Data = NULL,
  pvalSIG = 0.05,
  roll_median = 15
)

Arguments

AD_Data

Dataframe. Should at least contain chromosomal position ("pos"), overdispersion in controls and cases expressed as rho which lies between 0 and 1 ("rho_control", "rho_case") and the p-value testing for differential allelic divergence (difference in overdispersion parameter between controls and cases; "pval")

DE_Data

Dataframe. Should at least contain chromosomal position ("pos") and the log(2)FC of expression between cases and controls ("logFC")

Meth_Data

Dataframe. Optional; should at least contain chromosomal position ("pos") and the p-value testing for case hypermethylation ("logFC")

CNAgain_Data

Dataframe. Optional; should at least contain chromosomal position ("pos") and the average gain in cases ("CNAgain")

CNAloss_Data

Dataframe. Optional; should at least contain chromosomal position ("pos") and the average loss in cases ("CNAloss")

pvalSIG

Number. Significance cutoff of provided p-values (these should be FDR-corrected in advance)

roll_median

integer. All measures are plotted as rolling medians across roll_median data points.

allelefreq

Number. Allele frequency.

Value

A list containing the following components:

ADDE_plot

A plot vizualizing Allelic Divergence and Differential Expression results.

MethCNA_plot

A plot vizualizing hypermethylation and copy-number-alteration results.

LEG1, LEG2, LEG3

Various legend items, to be combined with ADDE_plot and MethCNA_plot if the user so desires, e.g. using the patchwork package.