Probability Mass Function of the binomial mixture distribution modeling population-level RNAseq data

pmf_binomMix calculates the probability of observing given population-level RNAseq data (i.e. both reference- and variant counts of one or more samples) assuming a binomial mixture model with parameter values as determined by the input. More specifically, the formula used is:

pr * dbinom(ref_counts, ref_counts + var_counts, prob = 1 - SE) +

pv * dbinom(var_counts, ref_counts + var_counts, prob = 1 - SE) +

prv * dbinom(ref_counts, ref_counts + var_counts, prob = probshift)

Usage

pmf_binomMix(ref_counts, var_counts, probshift, SE, pr, pv, prv)

Arguments

ref_counts

Number or Numeric vector Reference count(s).

var_counts

Number or Numeric vector. Variant count(s).

probshift

Number. The reference allele fraction in heterozygotes, indicating allelic bias when deviating from 0.5

SE

Number. Sequencing error rate.

pr

Number. Reference homozygote genotype probability of the locus.

pv

Number. Variant homozygote genotype probability of the locus.

prv

Number. Heterozygote genotype probability of the locus.

Value

Probability of observing ref_counts and var_counts

Examples

pmf_binomMix(10, 10, 0.5, 0.002, 0.25, 0.25, 0.5)
#> [1] 0.08809853
pmf_binomMix(0, 10, 0.8, 0.002, 0.25, 0.25, 0.5)
#> [1] 0.2450448
pmf_binomMix(c(5, 8, 10, 3, 5, 6, 23), c(8, 8, 6, 4, 4, 10, 0), 0.8, 0.002, 0.25, 0.25, 0.5)
#> [1] 0.0005398069 0.0027638115 0.0275152785 0.0143360696 0.0330301445
#> [6] 0.0001074816 0.2417009722
pmf_binomMix(c(5, 0, 0, 3, 5, 1, 23), c(1, 8, 6, 2, 0, 10, 0), 0.8, 0.002, 0.25, 0.25, 0.5)
#> [1] 0.199578120 0.246029168 0.247046960 0.102409960 0.411349980 0.005391435
#> [7] 0.241700972