Performs a differential Allelic Divergence analysis on a control- and case-dataset

dAD_analysis is a wrapper function bundling EMfit_betabinom_robust, EMfit_betabinom, EMfit_betabinom_popcomb, pmf_betabinomMix, and HWE_chisquared to run the entire differential Allelic Divergence detection analysis as discussed in the package vignette. To this end, it goes over the following steps:

Calling EMfit_betabinom_robust on both the control- and case-dataset separately for the sole purpose of outlier detection; detected outliers in both datasets play no part in future Expectation-Maximization-fits.

Perform one fit on the joint (case+control) data using EMfit_betabinom, i.e. assuming all population parameters are shared including the heterozygous overdispersion parameter, i.e. assuming no differential Allelic Divergence.

Perform a fit on the control- and case-data using EMfit_betabinom_popcomb, which assumes all population parameters are shared except for the heteroyzous overdispersion parameter, i.e. a fit accomodating differential Allelic Divergence

Statistically test for differential Allelic Divergence using a likelihood ratio test with one degree of freedom comparing the fits from the two previous points.

Filling out the results dataframe dAD_res.

Usage

dAD_analysis(
  datas,
  SE,
  inbr,
  allelefreq = 0.5,
  dltaco = 10^-6,
  HWE = FALSE,
  p_InitEst = FALSE,
  p_inits = c(1/3, 1/3, 1/3),
  ThetaInits = "moment",
  ReEstThetas = "moment",
  NoSplitHom = TRUE,
  NoSplitHet = TRUE,
  ResetThetaMin = 10^-10,
  ResetThetaMax = 10^-1,
  DistRob = "Cook",
  CookMargin = 5,
  LikEmpNum = 1000,
  LikMargin = 0,
  NumHetMin = 5,
  MaxOutFrac = 0.5,
  thetaTRY = c(10^-1, 10^-3, 10^-7),
  probshift_init = 0.5,
  FirstFewFixed = NULL,
  MemLim = 2048,
  Xtra = 5,
  epsabs = 0.001,
  MaxIt = 100
)

Arguments

SE

Number. Sequencing error rate population metaparameter.

inbr

Number. Inbreeding coefficient population metaparameter.

allelefreq, dltaco, HWE, p_InitEst, ThetaInits, ReEstThetas, NoSplitHom, NoSplitHet, ResetThetaMin, ResetThetaMax, DistRob, CookMargin, LikEmpNum, LikMargin, NumHetMin, MaxOutFrac, thetaTRY

Remaining control parameters to be passed to EMfit_betabinom_robust, EMfit_betabinom, and EMfit_betabinom_popcomb. See their respective documentation for more information. Usually, these parameter's default values are fine.

Datalist

A list containing two lists of dataframes, which in turn contain control- and case-data (in that order). Subsequent entries in both lists correspond to subsequent loci (lists must be the same size and named per-locus). Each of the dataframes should at least contain both a ref_count and var_count column.

Value

A dataframe with one row per locus, containing for that locus in its columns:

LocName

Locus name.

PiFitH0

Heterozygous pi parameter (reference allele fraction) for the joint fit sharing all parameters.

PiFitH1

Heterozygous pi parameter (reference allele fraction) for the fit allowing differential Allelic Divergence.

ThetaHetH0

Heterozygous theta parameter (overdispersion) for the joint fit sharing all parameters.

ThetaHetCTRL

Heterozygous theta parameter (overdispersion) of controls for the fit allowing differential Allelic Divergence.

ThetaHetCASE

Heterozygous theta parameter (overdispersion) of cases for the fit allowing differential Allelic Divergence.

RhoHetH0

Heterozygous rho parameter (overdispersion, ranging from 0 to 1) for the joint fit sharing all parameters.

RhoHetCTRL

Heterozygous rho parameter (overdispersion, ranging from 0 to 1) of controls for the fit allowing differential Allelic Divergence.

RhoHetCASE

Heterozygous rho parameter (overdispersion, ranging from 0 to 1) of cases for the fit allowing differential Allelic Divergence.

NumHetCTRL

Estimate number of heterozygotes in controls, according to its separate outlier-detection fit from step 1 above. Outliers are themselves genotyped and counted towards this number, just not used in the actual fitting procedure.

NumHetCASE

Estimate number of heterozygotes in cases, according to its separate outlier-detection fit from step 1 above. Outliers are themselves genotyped and counted towards this number, just not used in the actual fitting procedure.

RobFlagCTRL

The RobFlag output of the separate outlier-detection fit from step 1 on controls, see EMfit_betabinom_robust's documentation.

RobFlagCASE

The RobFlag output of the separate outlier-detection fit from step 1 on cases, see EMfit_betabinom_robust's documentation.

HWECTRL

HWE chi squared p-value for controls, see HWE_chisquared. Performed on genotyping data according to the outlier-detection fit of step 1.

HWECASE

HWE chi squared p-value for cases, see HWE_chisquared. Performed on genotyping data according to the outlier-detection fit of step 1.

CovCTRL_mean

Mean coverage (reference + variant count) across control data.

CovCASE_mean

Mean coverage (reference + variant count) across case data.

CovCTRL_med

Median coverage (reference + variant count) across control data.

CovCASE_med

Median coverage (reference + variant count) across case data.

NumOutCTRL

Number of outliers detected in control data.

NumOutCASE

Number of outliers detected in case data.

QualityCTRL

Quality flag returned by EMfit_betabinom_robust (see documentation) of the outlier-detection fit on control data (step 1 above).

QualityCASE

Quality flag returned by EMfit_betabinom_robust (see documentation) of the outlier-detection fit on case data (step 1 above).

pr

Reference homozygote fraction according to the differential Allelic Divergence accomodating fit (different heterozygous overdispersion parameters).

prv

Heterozygote fraction according to the differential Allelic Divergence accomodating fit (different heterozygous overdispersion parameters).

pv

Variant homozygote fraction according to the differential Allelic Divergence accomodating fit (different heterozygous overdispersion parameters).

ThetaHomH0

Homozygous theta parameter (overdispersion) for the joint fit sharing all parameters.

ThetaHomH1

Homozygous theta parameter (overdispersion) for the fit allowing differential Allelic Divergence.