Probability Mass Function of the beta-binomial mixture distribution modeling population-level RNAseq data

pmf_betabinomMix calculates the probability of observing given population-level RNAseq data (i.e. both reference- and variant counts of one or more samples) assuming a beta-binomial mixture model with parameter values as determined by the input. More specifically, the formula used is (using maelstRom's dBetaBinom function):

pr * dBetaBinom(ref_counts, ref_counts + var_counts, pi = 1 - SE, theta = theta_hom, LOG = FALSE) +

pv * dBetaBinom(var_counts, ref_counts + var_counts, pi = 1 - SE, theta = theta_hom, LOG = FALSE) +

prv * dBetaBinom(ref_counts, ref_counts + var_counts, pi = probshift, theta = theta_het, LOG = FALSE)

pmf_betabinomMix(
  ref_counts,
  var_counts,
  probshift,
  SE,
  pr,
  pv,
  prv,
  theta_hom,
  theta_het
)

Arguments

ref_counts

Number or Numeric vector Reference count(s).

var_counts

Number or Numeric vector. Variant count(s).

probshift

Number. The reference allele fraction in heterozygotes, indicating allelic bias when deviating from 0.5

SE

Number. Sequencing error rate.

pr

Number. Reference homozygote genotype probability of the locus.

pv

Number. Variant homozygote genotype probability of the locus.

prv

Number. Heterozygote genotype probability of the locus.

theta_hom

Number. The dispersion parameter of the homozygous peaks.

theta_het

Number. The dispersion parameter of the heterozygous peak.

Value

Probability of observing ref_counts and var_counts